using isAuthenticated rather than OAuth

+++ controllers/titles.js
@@ -4,7 +4,7 @@ const express = require('express');
 const router = express.Router();
 const axios = require('axios');
 
-const getAccessToken = () => {
+/*const getAccessToken = () => {
   return new Promise((resolve, reject) => {
     const options = {
       method: 'POST',
@@ -31,7 +31,7 @@ const getAccessToken = () => {
         reject(error);
       });
   });
-};
+};*/
 
 const getTitles = async (req, res) => {
   try {
@@ -140,22 +140,33 @@ DELETE route for deleting an author
 */
 const deleteTitle = async (req, res) => {
   try {
+    // check if the user is logged in with oidc.isAuthenticated()
+    // if not, redirect to the login page
+    if (!req.oidc.isAuthenticated()) {
+      return res.redirect('/login');
+    }
+
     const deleteTitleID = new ObjectId(req.params.id);
 
     // Get the access token
-    const accessToken = await getAccessToken();
+    //const accessToken = await getAccessToken();
 
     // Include the Authorization header with the access token
-    const headers = {
+    /*const headers = {
       Authorization: `Bearer ${accessToken}`,
-    };
+    };*/
 
-    // Make the DELETE request to your API endpoint
+    /*// Make the DELETE request to your API endpoint
     const response = await axios.delete(
       `https://yanezproject2library.onrender.com/titles/${deleteTitleID}`, {
         headers
       }
-    );
+    );*/
+
+    const result = await mongodb.getDb().db().collection('titles').deleteOne({
+      _id: deleteTitleID
+    });
+    console.log(result);
 
     // Check the response status and send the appropriate JSON response
     if (response.status === 200) {
@@ -176,7 +187,7 @@ const deleteTitle = async (req, res) => {
 };
 
 // Use the deleteTitle function as the DELETE route handler
-router.delete('/:id', deleteTitle);
+//router.delete('/:id', deleteTitle);
 
 
 module.exports = {
